+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+- Several binding fixes.
+
 ## 1.0.1
 
 - Update logo.